When asked to terminate, shutdown all active http connections and wait for their gophers to return, so that all tempfiles are removed.  Add test coverage.
Thanks to @tylerb for the done <- shutdown channel idea used in connection_tracker.go, from @stretchr's MIT `graceful` library.  Many other implementation ideas turned out to be similar to that library, but the implementation is different and has different behavior.

The biggest difference is that my implementation waits for the connection handlers to return even after their connection is shutdown, which was necessary to implement the required cleanup behavior.  This implementation is also factored separately to facilitate integration into the existing signal-handling infrastructure.  The wait/timeout behavior is however not implemented as not required/currently desired.